b+tree: Release v1 b+tree version
It is a time to release v1. The tree supports three operations
insert, delete and lookup. Insertion and deleting are done using
one way down-pass, without any recursions and any back updating.
Those are based on the preparatory operations(PO) during insert
or delete.

The stability test are passed with this v1 version.

Signed-off-by: Uladzislau Rezki (Sony) <urezki@gmail.com>